requirements.txt
Django
django-environ
django-precise-bbcode
django-simple-math-captcha
sorl-thumbnail
